Lighting seems to work better now. but! there's a problem with the rectangle normals
if a surface turns up completely black, it's because its normal is calculated wrong. I did a *very* quick fix on line 86 of rectangle that inverts the normal and it fixed some examples. maybe by luck this has solved all of our problems - but I think there should still be a deeper problem with the way the normal is calculated - he said in class it should depend on the order of p0, p1, p2, meaning the two vectors (p0->p1) and (p1->p2) and right now it does it using the vectors (p0->p1) and (p0->p2). fix if you can